// src/types/apps/PromoCodeType.ts
export type PromoCodeType = {
  id: number
  eName: string
  aName: string
  type: string
  discountType: string
  status: string
  brandName: string
  quantity: number
  percentageAmount: number
  minimumOrder: number
  maximumScope: number
  usagePerCustomer: number
  user: string
  startDate: string
  endDate: string
  channel: string
  products: string
  cities: string
}